refactor classname and packages, change AndroidManifest content
