+++ app/src/main/java/com/university/soa/bus/Book.java
@@ -6,7 +6,7 @@ import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 
-import com.university.soa.bus.SeatClass.TourSelection;
+import com.university.soa.bus.seatclass.TourSelection;
 
 import org.parceler.Parcels;
