+++ app/src/main/java/com/university/soa/bus/SavedSeats.java
@@ -32,7 +32,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import Models.AppStatus;
+import models.AppStatus;
 
 import static android.view.View.INVISIBLE;
 import static android.view.View.VISIBLE;
