+++ app/src/main/java/models/AppStatus.java
@@ -1,4 +1,4 @@
-package Models;
+package models;
 
 import android.content.Context;
 import android.net.ConnectivityManager;
